2021-ECP: made my pass on the tutorial
Convert the inline comments into markdown cells.  Expand the text in the
surrounding markdown cells to make self-guided use of the tutorial more
possible (and make the presentation day easier). Explicitly walkthrough
and callout the difference between the five different `flux mini`
sub-sub-commands.

Base the dockerfile off of the fluxrm/flux-sched image to simplify it.

Turns out that the issue with accessing the jupyter notebook came down
to having to specify `--ip 0.0.0.0` to the `jupyter notebook` command.